Fix formatting issues, refactor test
